Added package java7addons and moved all Classes, that require Java-7 there. If you now compile without this package you will get a Java6-Compatible version
Please note: There are now 2 InternalConfigDummy-Classes in existence:
In de.notepass.general.InternalConfig: This is now just the Configuration for the Java-6 package!
In de.notepass.general.java7addons.InternalConfig.InternalConfigDummy: This class is the configuration for the Java6 and Java7-package!

So, if you used Java7-Functionality, please import de.notepass.general.java7addons.InternalConfig.InternalConfigDummy from now on